requirements.txt
ffmpeg-python
colorthief
pillow
pytube
pedalboard
numpy
pydub